feat: enable multi-arch builds with ARM compatibility (#206)
Setup multi-arch builds with a manifest that will automatically point to the image of supported architectures.

On Kubernetes, with these changes, images are still referred to in the same manner, but will resolve to the architecture that is relevant to the node where it is to be run. 
```
                                                  --> quay.io/<user>/llama-stack-k8s-operator/<version>-linux-arm64
quay.io/<user>/llama-stack-k8s-operator/<version>-|
                                                  --> quay.io/<user>/llama-stack-k8s-operator/<version>-linux-amd64
```

This should allow users to seamlessly transition between clusters with different architectures, once the distribution changes are also in place.

Closes: RHAIENG-1941


Approved-by: mfleader

Approved-by: VaishnaviHire